+++ apps/mobile/src/components/posts/video/red-gifs.tsx
@@ -1,13 +1,10 @@
 import { View } from 'react-native'
-import { type SharedValue } from 'react-native-reanimated'
 import { StyleSheet } from 'react-native-unistyles'
 
 import { Spinner } from '~/components/common/spinner'
-import { useFocused } from '~/hooks/focus'
 import { useRedGifs } from '~/hooks/red-gifs'
 import { type PostMedia } from '~/types/post'
 
-import { VideoPlaceholder } from './placeholder'
 import { VideoPlayer } from './player'
 
 type Props = {
@@ -18,7 +15,6 @@ type Props = {
   recyclingKey: string
   spoiler?: boolean
   video: PostMedia
-  viewing?: SharedValue<string | null>
 }
 
 export function RedGifsVideo({
@@ -29,10 +25,7 @@ export function RedGifsVideo({
   recyclingKey,
   spoiler,
   video,
-  viewing,
 }: Props) {
-  const { focused } = useFocused()
-
   styles.useVariants({
     compact,
     crossPost,
@@ -40,20 +33,6 @@ export function RedGifsVideo({
   })
 
   const { gif } = useRedGifs(video.url)
-
-  if (!focused) {
-    return (
-      <VideoPlaceholder
-        compact={compact}
-        crossPost={crossPost}
-        large={large}
-        nsfw={nsfw}
-        recyclingKey={recyclingKey}
-        spoiler={spoiler}
-        video={video}
-      />
-    )
-  }
 
   if (gif) {
     return (
@@ -68,7 +47,6 @@ export function RedGifsVideo({
           ...video,
           url: gif.url,
         }}
-        viewing={viewing}
       />
     )
   }
@@ -114,7 +92,7 @@ const styles = StyleSheet.create((theme, runtime) => ({
       },
     ],
     justifyContent: 'center',
-    maxHeight: runtime.screen.height * 0.4,
+    maxHeight: runtime.screen.height * 0.5,
     overflow: 'hidden',
     variants: {
       compact: {